requirements.txt
# ── Web framework ─────────────────────────────────────
Flask==3.1.0
flask-cors==5.0.0
gunicorn==23.0.0

# ── Database ──────────────────────────────────────────
psycopg2-binary==2.9.10

# ── Auth ──────────────────────────────────────────────
PyJWT==2.10.1
bcrypt==4.2.1

# ── AI / HTTP ─────────────────────────────────────────
httpx==0.28.1
openai==1.59.6

# ── Utilities ─────────────────────────────────────────
python-dotenv==1.0.1
Werkzeug==3.1.3
